requirements.txt
pystray==0.19.5
pycaw==20240210
Pillow==10.2.0
comtypes==1.4.1
pywin32==306
feedparser==6.0.11
pyperclip==1.9.0
pygame==2.5.2
pytest>=7.4.4
flake8>=6.0.0
# 音訊處理和等化器相關套件
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
# CustomTkinter for modern UI
customtkinter>=5.2.0
# RAG assistant dependencies
chromadb>=0.4.0
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
# Discord Rich Presence
pypresence>=4.2.0
# YouTube Music API
ytmusicapi>=1.0.0
# 媒體鍵支持
pynput>=1.7.6
# Windows 媒體通知整合
winrt-Windows.Media>=2.0.0
# 音頻文件元數據編輯
mutagen>=1.47.0
requests
